Split json request data into api and ui folders to differentiate more easily
