+++ e2e/tests/manageCases/caseProgression/orders/C100/createAnOrder/judge/createBlankOrderOrDirectionsOrder.spec.ts
@@ -7,7 +7,7 @@ import { ManageOrder5Params } from "../../../../../../../pageObjects/pages/exui/
 import { OrderInformation } from "../../../../../../../pageObjects/pages/exui/caseView/draftOrders.po.ts";
 import { test } from "../../../../../../fixtures.ts";
 import config from "../../../../../../../utils/config.utils.ts";
-import { C21CreateOrderScenarios } from "../../../../../../../testData/manageOrders.ts";
+import { C21CreateOrderScenarios } from "../../../../../../../testData/ui/manageOrders.ts";
 import { ManageOrder4Params } from "../../../../../../../pageObjects/pages/exui/orders/manageOrders/manageOrder4.po.ts";
 import { ManageOrder30Params } from "../../../../../../../pageObjects/pages/exui/orders/manageOrders/manageOrder30.po.ts";
