+++ e2e/tests/manageCases/caseProgression/orders/C100/uploadAnOrder/uploadChildArrangementsOrderJudge.spec.ts
@@ -1,6 +1,6 @@
 import { test } from "../../../../../fixtures.ts";
 import config from "../../../../../../utils/config.utils.ts";
-import { C43UploadOrderScenarios } from "../../../../../../testData/manageOrders.ts";
+import { C43UploadOrderScenarios } from "../../../../../../testData/ui/manageOrders.ts";
 import {
   manageOrdersOptions,
   OrderTypes,
