+++ e2e/tests/manageCases/caseProgression/orders/FL401/createAnOrder/caseWorker/createAmendedDischargedOrVariedOrder.spec.ts
@@ -9,7 +9,7 @@ import { ManageOrder5Params } from "../../../../../../../pageObjects/pages/exui/
 import { ManageOrder19Params } from "../../../../../../../pageObjects/pages/exui/orders/manageOrders/manageOrder19.po.ts";
 import { ManageOrder24Params } from "../../../../../../../pageObjects/pages/exui/orders/manageOrders/manageOrder24.po.ts";
 import { OrderInformation } from "../../../../../../../pageObjects/pages/exui/caseView/Orders.po.ts";
-import { AmendedDischargedVariedOrderScenarios } from "../../../../../../../testData/manageOrders.ts";
+import { AmendedDischargedVariedOrderScenarios } from "../../../../../../../testData/ui/manageOrders.ts";
 import { ManageOrder26Params } from "../../../../../../../pageObjects/pages/exui/orders/manageOrders/manageOrder26.po.ts";
 import { ManageOrder28Params } from "../../../../../../../pageObjects/pages/exui/orders/manageOrders/manageOrder28.po.ts";
