+++ e2e/tests/manageCases/caseProgression/orders/FL401/createAnOrder/caseWorker/createOccupationOrder.spec.ts
@@ -10,7 +10,7 @@ import { ManageOrder12Params } from "../../../../../../../pageObjects/pages/exui
 import { ManageOrder19Params } from "../../../../../../../pageObjects/pages/exui/orders/manageOrders/manageOrder19.po.ts";
 import { ManageOrder24Params } from "../../../../../../../pageObjects/pages/exui/orders/manageOrders/manageOrder24.po.ts";
 import { OrderInformation } from "../../../../../../../pageObjects/pages/exui/caseView/draftOrders.po.ts";
-import { OccupationOrderScenarios } from "../../../../../../../testData/manageOrders.ts";
+import { OccupationOrderScenarios } from "../../../../../../../testData/ui/manageOrders.ts";
 
 export interface OccupationOrderParams {
   name: string;
