+++ e2e/tests/manageCases/caseProgression/orders/FL401/createAnOrder/solicitor/draftAnOrderNonMolestation.spec.ts
@@ -1,6 +1,6 @@
 import { test } from "../../../../../../fixtures.ts";
 import Config from "../../../../../../../utils/config.utils.ts";
-import { NonMolestationDraftOrderScenarios } from "../../../../../../../testData/draftOrders.ts";
+import { NonMolestationDraftOrderScenarios } from "../../../../../../../testData/ui/draftOrders.ts";
 import {
   OrderTypes,
   solicitorCaseCreateType,
